Revert "Moved and renamed vision command folder"
This reverts commit 1261071dc67e502e4cbd0eba918a33116df7fd40.
