+++ src/main/java/frc/robot/VisionCommands/aimToSpeaker.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.visioncommands;
+package frc.robot.VisionCommands;
 
 import java.util.function.DoubleSupplier;
